requirements.txt
Flask
Flask-sQLAlchemy
pytest
flasgger
openai
python-dotenv
requests
pydantic
mysql-connector-python
jinja2